requirements.txt
openai
chromadb
python-dotenv
faiss-cpu
numpy
tiktoken
langchain-text-splitters